Adding JWT-Extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 from models.store import StoreModel
 
@@ -40,13 +40,16 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
             return {'message': 'Item deleted'}
         return{'message': 'The item "{}" does not existe'.format(name)}, 404
     
-    @jwt_required()
+    @jwt_required(fresh=True)
     def put(self, name):
         data = Item.parse.parse_args()
         item = ItemModel.find_by_name(name)
@@ -66,6 +69,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    @jwt_required()
+    @jwt_required(optional=True, fresh=False)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.get_all_items()]}
+        user_id = get_jwt_identity()
+        items = ItemModel.get_all_items()
+        if user_id:
+            return {'items': [item.json() for item in items]}, 200
+        return {'items': [item.name for item in items],
+                'message': 'More data available if you log in'
+        }, 200
+
